+++ worlds/banjo_tooie/Options.py
@@ -215,6 +215,18 @@ class EnableHoneyBRewards(DefaultOnToggle):
     """Health Upgrades are added to the pool."""
     display_name = "Randomize Honey B Rewards"
 
+class RandomizeGreenRelics(DefaultOnToggle):
+    """Targitzan's Green Relics are randomized."""
+    display_name = "Randomize Green Relics"
+
+class RandomizeBeans(DefaultOnToggle):
+    """CCL Beans are randomized."""
+    display_name = "Randomize Beans"
+
+class RandomizeBigTentTickets(DefaultOnToggle):
+    """Big Top Tickets are randomized."""
+    display_name = "Randomize Big Top Tickets"
+
 class RandomizeGlowbos(DefaultOnToggle):
     """Mumbo and Humba Magic are in the pool and automatically unlocked when received.
     When disabled, collecting a Glowbo will give you either a Mumbo or Humba Magic."""
@@ -412,6 +424,10 @@ class Backdoors(Toggle):
     George is pre-dropped to make HFP -> JRL more accessible."""
     display_name = "Open Backdoors"
 
+class GIFrontdoor(Toggle):
+    """Opens Grunty's Industries frontdoor without requiring to get in first."""
+    display_name = "Open GI Frontdoor"
+
 class OpenHag1(DefaultOnToggle):
     """HAG 1 boss fight is opened when Cauldron Keep is opened, requiring fewer Jiggies to win."""
     display_name = "Open HAG 1"
@@ -423,6 +439,10 @@ class RandomizeWorldOrder(Toggle):
 class RandomizeWorldZones(Toggle):
     """The main entrance of each world will warp you to a random world."""
     display_name = "Randomize World Entrances"
+
+class RandomizeBossZones(Toggle):
+    """The entrance of each boss will warp you to a random boss."""
+    display_name = "Randomize Bosses"
 
 class RandomizeStopnSwap(Toggle):
     """Mystery Eggs, their rewards, and the Ice Key are scattered across the MultiWorld."""
@@ -641,6 +661,9 @@ class BanjoTooieOptions(PerGameCommonOptions):
     cheato_rewards: RandomizeCheatoRewards
     randomize_honeycombs: RandomizeHoneycombs
     honeyb_rewards: EnableHoneyBRewards
+    randomize_tickets:RandomizeBigTentTickets
+    randomize_green_relics:RandomizeGreenRelics
+    randomize_beans:RandomizeBeans
     randomize_glowbos: RandomizeGlowbos
     randomize_stop_n_swap: RandomizeStopnSwap
     randomize_dino_roar: RandomizeWorldDinoRoar
@@ -655,7 +678,9 @@ class BanjoTooieOptions(PerGameCommonOptions):
     skip_puzzles: SkipPuzzles
     randomize_worlds: RandomizeWorldOrder
     randomize_world_loading_zone: RandomizeWorldZones
+    randomize_boss_loading_zone: RandomizeBossZones
     backdoors: Backdoors
+    open_gi_frontdoor: GIFrontdoor
 
     signpost_hints: SignpostHints
     signpost_move_hints: SignpostMoveHints
@@ -730,6 +755,9 @@ bt_option_groups: List[OptionGroup] = [
         RandomizeCheatoRewards,
         RandomizeHoneycombs,
         EnableHoneyBRewards,
+        RandomizeBigTentTickets,
+        RandomizeGreenRelics,
+        RandomizeBeans,
         RandomizeGlowbos,
         RandomizeStopnSwap,
         RandomizeWorldDinoRoar,
@@ -746,7 +774,9 @@ bt_option_groups: List[OptionGroup] = [
         SkipPuzzles,
         RandomizeWorldOrder,
         RandomizeWorldZones,
+        RandomizeBossZones,
         Backdoors,
+        GIFrontdoor,
     ]),
 
     OptionGroup("Signpost Hints", [